requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.9.0
dnspython==2.4.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
langsmith==0.0.78
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.7.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pinecone-client==2.2.4
protobuf==4.25.1
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
streamlit==1.29.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
yarl==1.9.4
zipp==3.17.0
